tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.bezierCurveTo.ensuresubpath.1.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.bezierCurveTo.ensuresubpath.1</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.bezierCurveTo.ensuresubpath.1</h1>
<p class="desc">If there is no subpath, the first control point is added (and nothing is drawn up to it)</p>


<script>
var t = async_test("If there is no subpath, the first control point is added (and nothing is drawn up to it)");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);
  ctx.strokeStyle = '#f00';
  ctx.lineWidth = 50;
  ctx.beginPath();
  ctx.bezierCurveTo(100, 50, 200, 50, 200, 50);
  ctx.stroke();
  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixel(canvas, 95,45, 0,255,0,255);
  t.done();

});
</script>
